# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  description: REST API for managing bank cards with JWT authentication
  version: 1.0.0
  contact:
    name: Frolov Nikita Macsimovich
    email: nikitka_frolov_2014@inbox.ru
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/cards:
    get:
      tags: [Cards]
      summary: "Get user's cards"
      description: "Get paginated list of user's bank cards"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "owner"
      responses:
        '200':
          description: "List of user's cards"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardPageResponse'
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
  /api/cards/balance:
    get:
      tags: [Cards]
      summary: "Get card balance"
      description: "Get balance of specific card"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card balance"
          content:
            application/json:
              schema:
                type: number
                format: double
                example: 1500.50
        '404':
          description: "Card not found"
  /api/cards/activate:
    put:
      tags: [Cards]
      summary: "Activate card"
      description: "Activate user's card"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card activated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: "Get all users (Admin only)"
      description: "Get paginated list of all users (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "id"
      responses:
        '200':
          description: "List of all users"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/users/{userId}:
    get:
      tags: [Admin]
      summary: "Get user by ID (Admin only)"
      description: "Get user details by ID (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: "User details"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: "Admin access required"
        '404':
          description: "User not found"
  /api/admin/users/search:
    get:
      tags: [Admin]
      summary: "Search users by username (Admin only)"
      description: "Search users by username (case insensitive, Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
            example: "user"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "Search results"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/users/{userId}/role:
    put:
      tags: [Admin]
      summary: "Update user role (Admin only)"
      description: "Update user role (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            example: 2
        - name: newRole
          in: query
          required: true
          schema:
            type: string
            enum: [ROLE_USER, ROLE_ADMIN]
            example: "ROLE_ADMIN"
      responses:
        '200':
          description: "User role updated successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User role updated successfully"
                  userId:
                    type: integer
                    example: 2
                  newRole:
                    type: string
                    example: "ROLE_ADMIN"
        '403':
          description: "Admin access required"
  /api/admin/users/{userId}/block:
    put:
      tags: [Admin]
      summary: "Block user (Admin only)"
      description: "Block user account (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: "User blocked successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User blocked successfully"
                  userId:
                    type: integer
                    example: 2
                  username:
                    type: string
                    example: "user1"
        '403':
          description: "Admin access required"
  /api/admin/users/{userId}/activate:
    put:
      tags: [Admin]
      summary: "Activate user (Admin only)"
      description: "Activate user account (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: "User activated successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User activated successfully"
                  userId:
                    type: integer
                    example: 2
                  username:
                    type: string
                    example: "user1"
        '403':
          description: "Admin access required"
  /api/admin/users/stats:
    get:
      tags: [Admin]
      summary: "Get user statistics (Admin only)"
      description: "Get user statistics (total users, admin users, etc.)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "User statistics"
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUsers:
                    type: integer
                    example: 100
                  adminUsers:
                    type: integer
                    example: 5
                  regularUsers:
                    type: integer
                    example: 95
                  adminPercentage:
                    type: number
                    format: double
                    example: 5.0
        '403':
          description: "Admin access required"
  /api/admin/cards:
    get:
      tags: [Admin]
      summary: "Get all cards (Admin only)"
      description: "Get paginated list of all cards in system (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "List of all cards"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/users/{userId}/cards:
    get:
      tags: [Admin]
      summary: "Get user's cards (Admin only)"
      description: "Get paginated list of specific user's cards (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            example: 2
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "User's cards"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/block:
    put:
      tags: [Admin]
      summary: "Block any card (Admin only)"
      description: "Block any card in system (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card blocked successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/activate:
    put:
      tags: [Admin]
      summary: "Activate any card (Admin only)"
      description: "Activate any card in system (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card activated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/details:
    get:
      tags: [Admin]
      summary: "Get card details (Admin only)"
      description: "Get detailed card information (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
      responses:
        '200':
          description: "Card details"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/balance:
    put:
      tags: [Admin]
      summary: "Update card balance (Admin only)"
      description: "Update card balance (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
        - name: newBalance
          in: query
          required: true
          schema:
            type: number
            format: double
            minimum: 0
            example: 5000.00
      responses:
        '200':
          description: "Balance updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
  /api/admin/cards/by-status:
    get:
      tags: [Admin]
      summary: "Get cards by status (Admin only)"
      description: "Get paginated cards filtered by status (Admin only)"
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [ACTIVE, BLOCKED, EXPIRED]
            example: "ACTIVE"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "Filtered cards"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardPageResponse'
        '403':
          description: "Admin access required"

  # ========== NEW BLOCK REQUEST ENDPOINTS ==========
...
  /api/cards/block-requests:
    get:
      tags: [ Cards ]
      summary: "Get user's block requests"
      description: "Get paginated list of user's block requests"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "List of block requests"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestPageResponse'
  /api/admin/block-requests/pending:
    get:
      tags: [ Admin ]
      summary: "Get pending block requests (Admin only)"
      description: "Get paginated list of pending block requests (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "List of pending block requests"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/block-requests:
    get:
      tags: [ Admin ]
      summary: "Get all block requests (Admin only)"
      description: "Get paginated list of all block requests with optional status filter (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ PENDING, APPROVED, REJECTED ]
            example: "PENDING"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "List of block requests"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestPageResponse'
        '403':
          description: "Admin access required"
  /api/admin/block-requests/{requestId}/approve:
    put:
      tags: [ Admin ]
      summary: "Approve block request (Admin only)"
      description: "Approve block request and block the card (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: "Block request approved and card blocked"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
        '404':
          description: "Block request not found"
  /api/admin/block-requests/{requestId}/reject:
    put:
      tags: [ Admin ]
      summary: "Reject block request (Admin only)"
      description: "Reject block request and keep card active (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: reason
          in: query
          required: false
          schema:
            type: string
            example: "Insufficient reason provided"
      responses:
        '200':
          description: "Block request rejected and card remains active"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '403':
          description: "Admin access required"
        '404':
          description: "Block request not found"

  # ========== UPDATED CARD BLOCK ENDPOINT ==========
  /api/cards/block:
    put:
      tags: [ Cards ]
      summary: "Block card immediately (Admin only)"
      description: "Block card immediately without approval workflow (Admin only)"
      security:
        - bearerAuth: [ ]
      parameters:
        - name: cardNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^[0-9]{16}$'
            example: "4111111111111111"
      responses:
        '200':
          description: "Card blocked successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: "Admin access required"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 2
        username:
          type: string
          example: "user1"
        email:
          type: string
          example: "user1@bank.com"
        role:
          type: string
          enum: [ROLE_USER, ROLE_ADMIN]
          example: "ROLE_USER"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
...
    CardResponse:
      type: object
      properties:
        maskedNumber:
          type: string
          example: "**** **** **** 1111"
        owner:
          type: string
          example: "IVAN IVANOV"
        expiryDate:
          type: string
          format: date
          example: "2026-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          example: "ACTIVE"
        balance:
          type: number
          format: double
          example: 1500.50
        userId:
          type: integer
          example: 2
...
    CardPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalPages:
          type: integer
          example: 5
        totalElements:
          type: integer
          example: 42
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        last:
          type: boolean
          example: false
        first:
          type: boolean
          example: true
        numberOfElements:
          type: integer
          example: 10
        empty:
          type: boolean
          example: false
...
security:
  - bearerAuth: []